dev mode uses version 0.0.0

+++ tests/run_test.sh
@@ -11,23 +11,30 @@ VK_HEADERS_SANDBOX="${THIS_DIR}/../${VK_HEADERS_NAME}"
 
 test -d "${VK_HEADERS_SANDBOX}" && (echo "run gen_vku.sh first." 1>&2 && false)
 
-VK_HEADERS_TAG=$(
-  cd "${VK_HEADERS_SANDBOX}" &&
-  git describe --tag | grep -e "${VK_SDK_VER//\./\\.}" | sort -Vr | head -1
-  )
+if [[ "${1}" == "--dev" ]]; then
+  VK_VERSION="v0.0.0"
+else
+  VK_HEADERS_TAG=$(
+    cd "${VK_HEADERS_SANDBOX}" &&
+    git describe --tag | grep -e "${VK_SDK_VER//\./\\.}" | sort -Vr | head -1
+    )
 
-VK_VERSION="${VK_HEADERS_TAG//[^0-9.]/}"
-# prepend v and strip the fix number. most recent fix version should always prevail.
-VK_VERSION=$(set ${VK_VERSION//\./ }; echo "v${1}.${2}.${3}")
+  VK_VERSION="${VK_HEADERS_TAG//[^0-9.]/}"
+  # prepend v and strip the fix number. most recent fix version should always prevail.
+  VK_VERSION=$(set ${VK_VERSION//\./ }; echo "v${1}.${2}.${3}")
 
-if [[ -d build ]]; then
-  /bin/rm -rf build
+  if [[ -d build ]]; then
+    /bin/rm -rf build
+  fi
 fi
 
 set -eux
+
 mkdir -p build
 (cd build && cmake .. -DVK_VERSION="${VK_VERSION}")
+
 (cd build && cmake --build .)
+
 if [[ -d build/bin/Debug ]]; then
   TEST="./build/bin/Debug/test"
 else
